+++ CHANGELOG.md
@@ -8,5 +8,6 @@ All notable changes to **APK Lab** will be documented in this file.
 
 ### Added
 
-- DeAssemble/Decode APK file using `Apktool`
-- Rebuild APK file (Right click in `apktool.yml` file)
+- DeAssemble/Decode APK file using **`Apktool`**
+- Rebuild APK file (Context menu in `apktool.yml` file)
+- Sign the rebuilt APK using **`uber-apk-signer`**